[hotfix] [rebase] [rename] [fix] 프로젝트 계층 재구성 및 라우터 재설정
