+++ OmokFront/src/components/Footer/FooterBody/LeftContainer.vue
