+++ OmokFront/src/components/Footer/FooterBody/RightContainer.vue
