--- OmokFront/src/components/Header/HeaderMain/TitleAndCategory/Title/Title.vue
@@ -1,18 +1,0 @@
-<template>
-    <a href="guestMainPage">
-        <img class="image" src="@/assets/img/footer-title.png" alt="메인으로 가기" />
-    </a>
-</template>
-
-<script setup>
-
-</script>
-
-<style scoped>
-    .image {
-        position: relative;
-        width: 131px;
-        height: 58px;
-        object-fit: cover;
-    }
-</style>
