--- OmokFront/src/components/Header/HeaderMain/TitleAndCategory/TitleAndCategory.vue
@@ -1,21 +1,0 @@
-<template>
-    <div class="category-outter">
-        <Title/>
-        <Category/>
-    </div>
-</template>
-
-<script setup>
-import Category from './Category/Category.vue';
-import Title from './Title/Title.vue';
-</script>
-
-<style scoped>
-.category-outter {
-    display: inline-flex;
-    align-items: center;
-    gap: 39px;
-    position: relative;
-    flex: 0 0 auto;
-}
-</style>
