+++ OmokFront/src/main.js
@@ -1,6 +1,6 @@
 import { createApp } from 'vue'
 import App from './App.vue'
-import router from './router/router-index.js';
+import router from './router/router.js';
 
 const app = createApp(App);
 app.use(router);
