+++ OmokFront/src/views/GuestMainView.vue
